+++ src/BDSDetectorConstruction.cc
@@ -69,6 +69,9 @@ along with BDSIM.  If not, see <http://www.gnu.org/licenses/>.
 #include "G4VisAttributes.hh"
 #include "G4VPhysicalVolume.hh"
 #include "G4VSensitiveDetector.hh"
+#if G4VERSION_NUMBER > 1039
+#include "G4ChannelingOptrMultiParticleChangeCrossSection.hh"
+#endif
 
 #include "CLHEP/Units/SystemOfUnits.h"
 #include "CLHEP/Vector/EulerAngles.h"
@@ -768,7 +771,16 @@ void BDSDetectorConstruction::BuildPhysicsBias()
 
   G4String defaultBiasVacuum   = BDSParser::Instance()->GetOptions().defaultBiasVacuum;
   G4String defaultBiasMaterial = BDSParser::Instance()->GetOptions().defaultBiasMaterial;
-
+  
+  G4bool useDefaultBiasVacuum   = !defaultBiasVacuum.empty();
+  G4bool useDefaultBiasMaterial = !defaultBiasMaterial.empty();
+  const auto& biasObjectList = BDSParser::Instance()->GetBiasing();
+  G4bool biasesDefined = !biasObjectList.empty();
+  G4bool overallUseBiasing = useDefaultBiasVacuum || useDefaultBiasMaterial || biasesDefined;
+  G4cout << __METHOD_NAME__ << "Using generic biasing: " << BDS::BoolToString(overallUseBiasing) << G4endl;
+  if (!overallUseBiasing)
+    {return;} // no biasing used -> dont attach as just overhead for no reason
+  
   // apply per element biases
   for (auto const & item : *registry)
     {
@@ -790,7 +802,7 @@ void BDSDetectorConstruction::BuildPhysicsBias()
       
       // Build material bias object based on material bias list in the component
       auto egMaterial = BuildCrossSectionBias(accCom->GetBiasMaterialList(), defaultBiasMaterial, accName);
-      auto allLVs     = accCom->GetAllLogicalVolumes();
+      auto allLVs     = accCom->GetAllBiasingVolumes();
       if(debug)
 	{G4cout << __METHOD_NAME__ << "All logical volumes " << allLVs.size() << G4endl;}
       for (auto materialLV : allLVs)
@@ -805,6 +817,17 @@ void BDSDetectorConstruction::BuildPhysicsBias()
 	}
     }
   
+#if G4VERSION_NUMBER > 1039
+  // only available in 4.10.4 onwards
+  // crystal biasing necessary for implementation of variable density
+  std::set<G4LogicalVolume*>* crystals = BDSAcceleratorModel::Instance()->VolumeSet("crystals");
+  if (!crystals->empty())
+    {
+      auto crystalBiasing = new G4ChannelingOptrMultiParticleChangeCrossSection();
+      for (auto crystal : *crystals)
+	{crystalBiasing->AttachTo(crystal);}
+    }
+#endif
 #endif
 }